# TS: only transfer offsets as part of the AST